migrate fetch, bm25, hybrid queries to new dir structure
