+++ weaviate/collections/queries/bm25/__init__.py
@@ -1,0 +1,7 @@
+from .generate import _BM25Generate
+from .query import _BM25Query
+
+__all__ = [
+    "_BM25Generate",
+    "_BM25Query",
+]
