+++ weaviate/collections/queries/fetch_objects/__init__.py
@@ -1,0 +1,7 @@
+from .generate import _FetchObjectsGenerate
+from .query import _FetchObjectsQuery
+
+__all__ = [
+    "_FetchObjectsGenerate",
+    "_FetchObjectsQuery",
+]
