+++ weaviate/collections/queries/hybrid/__init__.py
@@ -1,0 +1,7 @@
+from .generate import _HybridGenerate
+from .query import _HybridQuery
+
+__all__ = [
+    "_HybridGenerate",
+    "_HybridQuery",
+]
