+++ weaviate/collections/queries/hybrid/query.pyi
@@ -1,0 +1,124 @@
+from typing import Generic, List, Literal, Optional, Type, overload
+
+from weaviate.collections.classes.filters import (
+    _Filters,
+)
+from weaviate.collections.classes.grpc import METADATA, PROPERTIES, REFERENCES, HybridFusion, Rerank
+from weaviate.collections.classes.internal import (
+    _QueryReturn,
+    References,
+    TReferences,
+    CrossReferences,
+)
+from weaviate.collections.classes.types import Properties, TProperties
+from weaviate.collections.queries.base import _BaseQuery
+
+class _HybridQuery(Generic[Properties, References], _BaseQuery[Properties, References]):
+    @overload
+    def hybrid(
+        self,
+        query: str,
+        *,
+        alpha: float = 0.5,
+        vector: Optional[List[float]] = None,
+        query_properties: Optional[List[str]] = None,
+        fusion_type: Optional[HybridFusion] = None,
+        limit: Optional[int] = None,
+        auto_limit: Optional[int] = None,
+        filters: Optional[_Filters] = None,
+        rerank: Optional[Rerank] = None,
+        include_vector: bool = False,
+        return_metadata: Optional[METADATA] = None,
+        return_properties: Optional[PROPERTIES] = None,
+        return_references: Literal[None] = None,
+    ) -> _QueryReturn[Properties, References]: ...
+    @overload
+    def hybrid(
+        self,
+        query: str,
+        *,
+        alpha: float = 0.5,
+        vector: Optional[List[float]] = None,
+        query_properties: Optional[List[str]] = None,
+        fusion_type: Optional[HybridFusion] = None,
+        limit: Optional[int] = None,
+        auto_limit: Optional[int] = None,
+        filters: Optional[_Filters] = None,
+        rerank: Optional[Rerank] = None,
+        include_vector: bool = False,
+        return_metadata: Optional[METADATA] = None,
+        return_properties: Optional[PROPERTIES] = None,
+        return_references: REFERENCES,
+    ) -> _QueryReturn[Properties, CrossReferences]: ...
+    @overload
+    def hybrid(
+        self,
+        query: str,
+        *,
+        alpha: float = 0.5,
+        vector: Optional[List[float]] = None,
+        query_properties: Optional[List[str]] = None,
+        fusion_type: Optional[HybridFusion] = None,
+        limit: Optional[int] = None,
+        auto_limit: Optional[int] = None,
+        filters: Optional[_Filters] = None,
+        rerank: Optional[Rerank] = None,
+        include_vector: bool = False,
+        return_metadata: Optional[METADATA] = None,
+        return_properties: Optional[PROPERTIES] = None,
+        return_references: Type[TReferences],
+    ) -> _QueryReturn[Properties, TReferences]: ...
+    @overload
+    def hybrid(
+        self,
+        query: str,
+        *,
+        alpha: float = 0.5,
+        vector: Optional[List[float]] = None,
+        query_properties: Optional[List[str]] = None,
+        fusion_type: Optional[HybridFusion] = None,
+        limit: Optional[int] = None,
+        auto_limit: Optional[int] = None,
+        filters: Optional[_Filters] = None,
+        rerank: Optional[Rerank] = None,
+        include_vector: bool = False,
+        return_metadata: Optional[METADATA] = None,
+        return_properties: Type[TProperties],
+        return_references: Literal[None] = None,
+    ) -> _QueryReturn[TProperties, References]: ...
+    @overload
+    def hybrid(
+        self,
+        query: str,
+        *,
+        alpha: float = 0.5,
+        vector: Optional[List[float]] = None,
+        query_properties: Optional[List[str]] = None,
+        fusion_type: Optional[HybridFusion] = None,
+        limit: Optional[int] = None,
+        auto_limit: Optional[int] = None,
+        filters: Optional[_Filters] = None,
+        rerank: Optional[Rerank] = None,
+        include_vector: bool = False,
+        return_metadata: Optional[METADATA] = None,
+        return_properties: Type[TProperties],
+        return_references: REFERENCES,
+    ) -> _QueryReturn[TProperties, CrossReferences]: ...
+    @overload
+    def hybrid(
+        self,
+        query: str,
+        *,
+        alpha: float = 0.5,
+        vector: Optional[List[float]] = None,
+        query_properties: Optional[List[str]] = None,
+        fusion_type: Optional[HybridFusion] = None,
+        limit: Optional[int] = None,
+        auto_limit: Optional[int] = None,
+        filters: Optional[_Filters] = None,
+        rerank: Optional[Rerank] = None,
+        include_vector: bool = False,
+        return_metadata: Optional[METADATA] = None,
+        return_properties: Type[TProperties],
+        return_references: Type[TReferences],
+    ) -> _QueryReturn[TProperties, TReferences]: ...
